inline module tests into their respective modules
